Make mypy pass: the annotations the refactor left behind
The lint workflow has run `mypy ./cardioception/ --ignore-missing-imports`
since before this branch, but the handoff's local lint recipe never listed it
and mypy was not installed in the venv, so nothing on this branch was ever
type checked. The first push found 22 errors. Nearly all are annotations that
the Phase 5 refactors made stale rather than anything the code does wrong:

- `trial()` still declared the 18-element tuple that `TrialOutcome` replaced
  in R8, so every attribute the caller reads off the result was an error, and
  its docstring still enumerated the tuple's members. Both now name
  `TrialOutcome`, which is where the column names already live.
- `responseDecision()` declared six return values and returns five. The
  docstring already said five; only the annotation was wrong.
- `listen_to_heart()` narrows on `accepted`, which mypy cannot connect back to
  `listenBPM`. It now narrows on the values themselves, which also makes the
  fallback visibly total.
- `recordedAt` is `Optional` because the exteroceptive path never records. The
  branch that uses it is the interoceptive one, so the invariant holds; it is
  now a `cast` with the reason written down rather than an implicit assumption.
- `marker_start` accepts a fractional starting position and was typed `int`.
- `append_silence` is called with fractional milliseconds by design: the
  inter-beat interval is rarely whole, and rounding it would change every file
  the generator produces. Widened to `float` rather than rounding the call.

No behaviour changes. 109 tests still pass, and black, isort and flake8 are
unchanged.

+++ cardioception/_rating.py
@@ -26,7 +26,7 @@ def keyboard_rating(
     low: int,
     high: int,
     labels: Sequence[str],
-    marker_start: Optional[int] = None,
+    marker_start: Optional[float] = None,
     min_time: float = 0.0,
     max_time: Optional[float] = None,
     accept_keys: Sequence[str] = ("down",),
